# inventario-api/docs/swagger.yaml
openapi: 3.0.3
info:
  title: API Inventario
  description: Documentación de la API de Inventario con ejemplos de uso
  version: 1.0.0
  contact:
    name: Soporte API
    email: soporte@empresa.com
servers:
  - url: http://localhost:3000
    description: Servidor local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Admin"
        email:
          type: string
          format: email
          example: "juan.admin@empresa.com"
        role:
          type: string
          enum: [admin, client]
          example: "admin"
        activo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-20T10:30:00.000Z"
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        numero_lote:
          type: string
          example: "LAPTOP001"
        nombre:
          type: string
          example: "MacBook Air M2"
        precio:
          type: number
          format: decimal
          example: 1299.99
        cantidad_disponible:
          type: integer
          example: 5
        descripcion:
          type: string
          example: "MacBook Air con chip M2, 8GB RAM, 256GB SSD"
        activo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: string
    RegisterRequest:
      type: object
      required:
        - nombre
        - email
        - password
      properties:
        nombre:
          type: string
          example: "Juan Admin"
        email:
          type: string
          format: email
          example: "juan.admin@empresa.com"
        password:
          type: string
          minLength: 6
          example: "password123"
        role:
          type: string
          enum: [admin, client]
          example: "admin"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "juan.admin@empresa.com"
        password:
          type: string
          example: "password123"
    ProductRequest:
      type: object
      required:
        - numero_lote
        - nombre
        - precio
        - cantidad_disponible
      properties:
        numero_lote:
          type: string
          example: "LAPTOP001"
        nombre:
          type: string
          example: "MacBook Air M2"
        precio:
          type: number
          format: decimal
          example: 1299.99
        cantidad_disponible:
          type: integer
          minimum: 0
          example: 5
        descripcion:
          type: string
          example: "MacBook Air con chip M2, 8GB RAM, 256GB SSD"
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Autenticación
      summary: Registrar nuevo usuario
      description: Registra un nuevo usuario en el sistema (admin o cliente)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              admin:
                summary: Registrar Admin
                value:
                  nombre: "Juan Admin"
                  email: "juan.admin@empresa.com"
                  password: "password123"
                  role: "admin"
              cliente:
                summary: Registrar Cliente
                value:
                  nombre: "María Cliente"
                  email: "maria@gmail.com"
                  password: "123456"
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: Autentica un usuario y retorna un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Productos
  /api/products:
    get:
      tags:
        - Productos
      summary: Listar productos
      description: Obtiene una lista paginada de productos con filtros opcionales
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número de página
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Elementos por página
        - name: search
          in: query
          schema:
            type: string
          description: Término de búsqueda
      responses:
        '200':
          description: Lista de productos obtenida exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          products:
                            type: array
                            items:
                              $ref: '#/components/schemas/Product'
                          pagination:
                            type: object
                            properties:
                              page:
                                type: integer
                              limit:
                                type: integer
                              total:
                                type: integer
                              pages:
                                type: integer

    post:
      tags:
        - Productos
      summary: Crear producto
      description: Crea un nuevo producto (solo administradores)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          description: Token faltante o inválido
        '403':
          description: Sin permisos de administrador
        '400':
          description: Error de validación
  /api/products/{id}:
    get:
      tags:
        - Productos
      summary: Obtener producto por ID
      description: Obtiene los detalles de un producto específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del producto
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado

    put:
      tags:
        - Productos
      summary: Actualizar producto
      description: Actualiza un producto existente (solo administradores)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                numero_lote:
                  type: string
                nombre:
                  type: string
                precio:
                  type: number
                  format: decimal
                cantidad_disponible:
                  type: integer
                descripcion:
                  type: string
              example:
                precio: 1199.99
                cantidad_disponible: 8
      responses:
        '200':
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          description: Token faltante o inválido
        '403':
          description: Sin permisos de administrador
        '404':
          description: Producto no encontrado

    delete:
      tags:
        - Productos
      summary: Eliminar producto
      description: Elimina un producto (soft delete - solo administradores)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del producto
      responses:
        '200':
          description: Producto eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Token faltante o inválido
        '403':
          description: Sin permisos de administrador
        '404':
          description: Producto no encontrado

  # Compras
...